backend/pyproject.toml
[tool.poetry]
name = "hackathon-todo-backend"
version = "0.1.0"
description = "Backend for the Hackathon Todo application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlmodel = "^0.0.22"
pydantic = "^2.10.1"
pydantic-settings = "^2.7.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true